MAHOUT-840 Build and Test regression forests
git-svn-id: https://svn.apache.org/repos/asf/mahout/trunk@1213034 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mahout/classifier/df/node/CategoricalNode.java
@@ -42,7 +42,7 @@ public class CategoricalNode extends Node {
   }
   
   @Override
-  public int classify(Instance instance) {
+  public double classify(Instance instance) {
     int index = ArrayUtils.indexOf(values, instance.get(attr));
     if (index == -1) {
       // value not available, we cannot predict
